requirements.txt
streamlit
pandas
faiss-cpu
numpy
sentence-transformers
